# backend/src/main/resources/openapi.yml
openapi: 3.0.0
info:
  title: API记录系统
  description: 用于记录和管理API调用信息的系统
  version: 1.0.0
  contact:
    name: 开发团队
    email: dev@example.com
servers:
  - url: http://localhost:8080
    description: 开发环境
  - url: http://172.16.2.95:8080
    description: 生产环境
paths:
...
  /users:
    get:
      summary: 获取所有用户
      description: 获取系统中的所有用户列表
      tags:
        - 用户管理
      security:
        - BasicAuth: []
      responses:
        '200':
          description: 成功获取用户列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserDto'
        '401':
          description: 未认证
        '403':
          description: 权限不足

    post:
      summary: 创建新用户
      description: 创建新的用户账户
      tags:
        - 用户管理
      security:
        - BasicAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateDto'
      responses:
        '201':
          description: 用户创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '400':
          description: 请求参数错误
        '401':
          description: 未认证
        '403':
          description: 权限不足
  /users/{id}:
    get:
      summary: 根据ID获取用户
      description: 根据用户ID获取用户详细信息
      tags:
        - 用户管理
      security:
        - BasicAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: 用户ID
      responses:
        '200':
          description: 成功获取用户信息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '404':
          description: 用户不存在
        '401':
          description: 未认证

    delete:
      summary: 删除用户
      description: 根据用户ID删除用户
      tags:
        - 用户管理
      security:
        - BasicAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: 用户ID
      responses:
        '200':
          description: 用户删除成功
        '404':
          description: 用户不存在
        '401':
          description: 未认证
        '403':
          description: 权限不足
  /users/{id}/update:
    post:
      summary: 更新用户信息
      description: 更新指定用户的信息
      tags:
        - 用户管理
      security:
        - BasicAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: 用户ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateDto'
      responses:
        '200':
          description: 用户信息更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '400':
          description: 请求参数错误
        '404':
          description: 用户不存在
        '401':
          description: 未认证
        '403':
          description: 权限不足
  /users/{id}/toggle-status:
    post:
      summary: 切换用户状态
      description: 启用或禁用用户账户
      tags:
        - 用户管理
      security:
        - BasicAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: 用户ID
      responses:
        '200':
          description: 用户状态切换成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '404':
          description: 用户不存在
        '401':
          description: 未认证
        '403':
          description: 权限不足
  /users/{id}/change-password:
    post:
      summary: 修改用户密码
      description: 修改指定用户的密码
      tags:
        - 用户管理
      security:
        - BasicAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: 用户ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - password
              properties:
                password:
                  type: string
                  description: 新密码
                  minLength: 6
      responses:
        '200':
          description: 密码修改成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '400':
          description: 请求参数错误
        '404':
          description: 用户不存在
        '401':
          description: 未认证
        '403':
          description: 权限不足
  /users/{id}/reset-password:
    post:
      summary: 重置用户密码
      description: 重置指定用户的密码为默认密码
      tags:
        - 用户管理
      security:
        - BasicAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: 用户ID
      responses:
        '200':
          description: 密码重置成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '404':
          description: 用户不存在
        '401':
          description: 未认证
        '403':
          description: 权限不足
  /users/stats:
    get:
      summary: 获取用户统计信息
      description: 获取用户总数、启用用户数、禁用用户数等统计信息
      tags:
        - 用户管理
      security:
        - BasicAuth: []
      responses:
        '200':
          description: 成功获取统计信息
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalUsers:
                    type: integer
                    description: 用户总数
                  enabledUsers:
                    type: integer
                    description: 启用用户数
                  disabledUsers:
                    type: integer
                    description: 禁用用户数
        '401':
          description: 未认证
        '403':
          description: 权限不足
  /users/generate-test-users:
    post:
      summary: 生成测试用户
      description: 生成指定数量的测试用户数据
      tags:
        - 用户管理
      security:
        - BasicAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - count
              properties:
                count:
                  type: integer
                  minimum: 1
                  maximum: 100
                  description: 要生成的用户数量
                  example: 10
      responses:
        '200':
          description: 测试用户生成成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "成功生成 10 个测试用户"
                  generatedUsers:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserDto'
                  count:
                    type: integer
                    example: 10
        '400':
          description: 请求参数错误
        '401':
          description: 未认证
        '403':
          description: 权限不足
...
components:
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
  schemas:
    UserDto:
      type: object
      description: 用户数据传输对象
      properties:
        id:
          type: integer
          description: 用户ID
        username:
          type: string
          description: 用户名
          example: "admin"
        email:
          type: string
          format: email
          description: 邮箱地址
          example: "admin@example.com"
        fullName:
          type: string
          description: 用户全名
          example: "系统管理员"
        phone:
          type: string
          description: 手机号码
          example: "13800138000"
        idCard:
          type: string
          description: 身份证号码
          example: "110101199001011234"
        department:
          type: string
          description: 部门
          example: "技术部"
        gender:
          type: string
          enum: [MALE, FEMALE, OTHER]
          description: 性别
          example: "MALE"
        officeAddress:
          type: string
          description: 办公地址
          example: "北京市朝阳区建国门外大街1号"
        bloodType:
          type: string
          enum: [A, B, AB, O]
          description: 血型
          example: "A"
        licensePlate:
          type: string
          description: 车牌号码
          example: "京A12345"
        homeAddress:
          type: string
          description: 住址
          example: "北京市海淀区中关村大街27号"
        landline:
          type: string
          description: 座机号码
          example: "010-12345678"
        enabled:
          type: boolean
          description: 是否启用
          example: true
        createdAt:
          type: string
          format: date-time
          description: 创建时间
        updatedAt:
          type: string
          format: date-time
          description: 更新时间
    UserCreateDto:
      type: object
      description: 创建用户请求对象
      required:
        - username
        - password
        - fullName
      properties:
        username:
          type: string
          description: 用户名
          minLength: 3
          maxLength: 20
          example: "newuser"
        password:
          type: string
          description: 密码
          minLength: 6
          example: "password123"
        fullName:
          type: string
          description: 用户全名
          example: "新用户"
        email:
          type: string
          format: email
          description: 邮箱地址
          example: "newuser@example.com"
        phone:
          type: string
          description: 手机号码
          example: "13800138000"
        idCard:
          type: string
          description: 身份证号码
          example: "110101199001011234"
        department:
          type: string
          description: 部门
          example: "技术部"
        gender:
          type: string
          enum: [MALE, FEMALE, OTHER]
          description: 性别
          example: "MALE"
        officeAddress:
          type: string
          description: 办公地址
          example: "北京市朝阳区建国门外大街1号"
        bloodType:
          type: string
          enum: [A, B, AB, O]
          description: 血型
          example: "A"
        licensePlate:
          type: string
          description: 车牌号码
          example: "京A12345"
        homeAddress:
          type: string
          description: 住址
          example: "北京市海淀区中关村大街27号"
        landline:
          type: string
          description: 座机号码
          example: "010-12345678"
        enabled:
          type: boolean
          description: 是否启用
          default: true
    UserUpdateDto:
      type: object
      description: 更新用户请求对象
      properties:
        fullName:
          type: string
          description: 用户全名
        email:
          type: string
          format: email
          description: 邮箱地址
        phone:
          type: string
          description: 手机号码
        idCard:
          type: string
          description: 身份证号码
        department:
          type: string
          description: 部门
        gender:
          type: string
          enum: [MALE, FEMALE, OTHER]
          description: 性别
        officeAddress:
          type: string
          description: 办公地址
        bloodType:
          type: string
          enum: [A, B, AB, O]
          description: 血型
        licensePlate:
          type: string
          description: 车牌号码
        homeAddress:
          type: string
          description: 住址
        landline:
          type: string
          description: 座机号码
        enabled:
          type: boolean
          description: 是否启用
...